+++ Src/AssemblyInfo.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyProduct("SplitPipeline")]
-[assembly: AssemblyVersion("1.0.0")]
+[assembly: AssemblyVersion("1.0.1")]
 [assembly: AssemblyTitle("SplitPipeline - Parallel Data Processing in PowerShell")]
 [assembly: AssemblyCompany("https://github.com/nightroman/SplitPipeline")]
 [assembly: AssemblyCopyright("Copyright (c) 2011-2013 Roman Kuzmin")]